IPFS Fetch Content Async (#15)
This makes the program faster (I hope), but also required quite a few changes. It did lead me to discover that we weren't very string with our handling of app data content (we were manipulating dictionaries). We also introduce a data structure for the Found and not found app data types.