[CIP-20] Update Query & Test (#30)
This PR finalizes the DB query based on the progress made in [rewards script](https://github.com/cowprotocol/solver-rewards) -- specifically that query which was finalized [here](https://github.com/cowprotocol/solver-rewards/pull/203)

Its a bit gross that the orderbook query lives in two places (for the time being). At some point we will want to expose this as an API endpoint (on a constrained `block_range`) directly via the orderbook. However, until we have landed on the correct query (sufficient to reproduce solver rewards) we will maintain this externally - in two places.

The tests introduced here are both for the DB extraction result and then the transformation phase to JSON before upload.

Closes #31