use more optimistic block range for sync
This PR addresses issue #72. At the moment the block range is computed
using the minimum of the latest finalized block from production and
staging and subtracting `REORG_THRESHOLD = 65`. This was done as a hot fix in #71.
One problem with this approach is that having few settlements in staging
blocks upload of production data to dune.

This PR changes the behavior to look at the maximum block of the latest
finalized blocks from production and staging subtracting
`MAX_PROCESSING_DELAY = 1`. This should leave enough time for the autopilot
to process settlements after finalization in staging and production.