[Mirror] Fix network fees (#111)
This PR mirrors the implementation of
https://github.com/cowprotocol/solver-rewards/pull/400

It should be adapted when the original implementation if changed. It
needs to be merged, hotfixed into prod, and data for the accounting week
starting on 2024-09-17 needs to be resynced. Otherwise the dashboard for
the next payment will be out of sync with solver-rewards.