Remove assert for solver overlap (#119)
This PR removes an assert which checked that solvers do only appear in
one of prod and barn. Instead, a warning is logged.

Even though in most cases a solver appearing in prod and barn is a sign
of invalid solver behavior, the rest of the code of this repo does not
depend on this assumption. Data is just concatenated and uploaded as is.

The [main rewards
query](https://github.com/cowprotocol/dune-queries/blob/main/cowprotocol/accounting/rewards/mainnet/mainnet_dashboard_query_2510345.sql)
on dune also does not depend on associating settlements to environments.
The environment is only added after aggregating payments
[here](https://github.com/cowprotocol/dune-queries/blob/6be6c1cfaa78b3dfc2ed65494c2cc0ef5c303889/cowprotocol/accounting/rewards/mainnet/mainnet_dashboard_query_2510345.sql#L118)
for easier digestion of the final table.