Sync price feed (#125)
This PR adds a simple job that pushes all prices collected in the
`prices` table of the analytics db to Dune. It mimics PR #103, and is
meant to run as a cronjob. I followed the "legacy method" as there are
some concerns about the current way Prefect has been integrated in this
repo.

I also prepared this PR in the infra repo.
https://github.com/cowprotocol/infrastructure/pull/2217

Note: there are currently not any tests. I saw there are some tests for
the app-data sync, so i could try to basically copy-paste those, as i am
not sure what's the best way to test this